starfish-ws/requirements.txt
Django==1.8.9
django-oauth-toolkit==0.10.0
djangorestframework==2.3.14
django-cors-headers==1.0.0
django-sendfile==0.3.6
jsonfield==0.9.22
python-dateutil==2.1
pytz==2015.4
kazoo==2.0
twilio==4.0.0
gunicorn==19.3.0
Pillow==2.7.0
audioread==1.2.1
python-magic==0.4.6
python3-protobuf==2.5.0
simpleflake==0.1.5
twisted==14.0.2
requests==2.6.0
python3-memcached==1.51
httpagentparser==1.7.6
phonenumbers==7.0.4
dj-static==0.0.6
XlsxWriter==0.6.7
psycopg2==2.6
pycrypto==2.6.1
elasticsearch==1.6.0
dnspython3==1.12.0
django-bitfield==1.8.0
pinyin==0.2.5
psd-tools==1.2
python3-pika==0.9.14
oauthlib==1.0.3
lz4==0.7.0
git+https://github.com/simon-liu/apns-client
git+https://github.com/simon-liu/kafka-python
uwsgi==2.0.12
